+++ src/Symfony/Component/Messenger/Bridge/Beanstalkd/CHANGELOG.md
@@ -1,6 +1,12 @@
 CHANGELOG
 =========
 
+7.1
+---
+
+ * Add `bury_on_reject` option to bury failed messages instead of deleting them
+ * Add `BeanstalkdPriorityStamp` option to allow setting the message priority
+
 5.2.0
 -----